Corrected Hibernate mappings. Sorted out default namespaces in parsers.
git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@3501 7c6358e6-4a41-0410-a743-a5b2a554c398

+++ src/org/biojavax/SimpleDocRef.java
@@ -237,13 +237,13 @@ public class SimpleDocRef extends AbstractChangeable implements DocRef {
     public String toString() { return this.getAuthors()+"; "+this.getLocation(); }
     
     // Hibernate requirement - not for public use.
-    private Long id;
+    private Integer id;
     
     // Hibernate requirement - not for public use.
-    private Long getId() { return this.id; }
+    private Integer getId() { return this.id; }
     
     // Hibernate requirement - not for public use.
-    private void setId(Long id) { this.id = id; }
+    private void setId(Integer id) { this.id = id; }
     
 }
