+++ src/org/biojavax/bio/SimpleBioEntryRelationship.java
@@ -153,12 +153,12 @@ public class SimpleBioEntryRelationship extends AbstractChangeable implements Bi
     public String toString() { return this.getTerm()+"("+this.getSubject()+")"; }
     
     // Hibernate requirement - not for public use.
-    private Long id;
+    private Integer id;
     
     // Hibernate requirement - not for public use.
-    private Long getId() { return this.id; }
+    private Integer getId() { return this.id; }
     
     // Hibernate requirement - not for public use.
-    private void setId(Long id) { this.id = id; }
+    private void setId(Integer id) { this.id = id; }
 }
