+++ src/org/biojavax/bio/taxa/SimpleNCBITaxon.java
@@ -381,13 +381,13 @@ public class SimpleNCBITaxon extends AbstractChangeable implements NCBITaxon {
     public String toString() { return "taxid:"+this.getNCBITaxID(); }
     
     // Hibernate requirement - not for public use.
-    private Long id;
+    private Integer id;
     
     // Hibernate requirement - not for public use.
-    private Long getId() { return this.id; }
+    private Integer getId() { return this.id; }
     
     // Hibernate requirement - not for public use.
-    private void setId(Long id) { this.id = id; }
+    private void setId(Integer id) { this.id = id; }
     
 }
