+++ src/org/biojavax/ontology/SimpleComparableTriple.java
@@ -275,13 +275,13 @@ public class SimpleComparableTriple extends AbstractChangeable implements Compar
     
     
     // Hibernate requirement - not for public use.
-    private Long id;
+    private Integer id;
     
     
     // Hibernate requirement - not for public use.
-    private Long getId() { return this.id; }
+    private Integer getId() { return this.id; }
     
     
     // Hibernate requirement - not for public use.
-    private void setId(Long id) { this.id = id; }
+    private void setId(Integer id) { this.id = id; }
 }
